Add actionbar to Dialer
Initial changes to replace the fake actionbar affordances in Dialer
with an actual actionBar. This puts the dialer in an intermediate
hybrid state where we are now using the system actionbar for search,
but the searchview is not yet themed to look like our old search
bar, so this is definitely still WIP UI.

This CL retains the fake action bar at the bottom of the screen to
continue allowing quick access to call history while other changes
are underway, but replaces the searchview and the overflow menu
with framework provided actionbar functionality.

Other related changes:
* The searchview (in the top actionbar) is hidden when scrolling to
devote the entire screen area to the speeddial list
https://folio.googleplex.com/phone/L/0407/#/03.png
* Functionality provided by the fake overflow menu has been moved to
 the real actionbar overflow menu
* Various modifications to child fragments/listviews to take into
account the existence of an actionbar
* Cleaned up unneeded styles and related-dead code

Bug: 13933082

Change-Id: Ifd855bd67af7bf2d2a403ef9556ade54e8d57ee6