Use ViewPager in main Dialer view
This change replaces PhoneFavoriteFragment with ListsFragment,
a fragment that contains a Viewpager that will eventually
contain 3 fragments - Speed Dial, Recents (an abridged call log) and
All Contacts. For now, only speed dial and all contacts are in the
viewpager due to the call log fragment not playing nice with being
embedded in a parent fragment.

ViewPagerTabs is a newly added custom view that serves as a indicator
for ViewPager tabs. It behaves similarly to the newly deprecated
ActionBar tabs, but can be placed anywhere on screen.

Bug: 13935070

Change-Id: I916c516dc295246b2a95de2f0dc726784c2ee0cc