Reveal recent calls shortcut only when at top of list
Add an intermediate offset to OverlappingPaneLayout so that the
search box can be exposed without revealing the shortcut card as well.

Add state tracking on OverlappingPaneLayout so that when a list is in
the middle of a scroll, the shortcut card is not revealed. Only when
dragging the list downwards after the top of the list has been reached,
should the shortcut card be revealed.

Bug: 15407905

Change-Id: Ie303f4a36ce14f2b1103865f5583e3cab6263deb