Handle searchbox long press to speed up copy/paste.
Before, user needed to first tap on the the collapsed search box to
expand it. Afterwards, needed to long-press to open the paste option.

Now, if the user long-presses on the collapsed search box, it
simulates a click on the search box (causing it to expand), and then
invokes a long-press to open the paste option.

Bug: 17882004
Change-Id: Ie23df56fe1c340aacf484bc482fa8abb06a3574c