Refactor ActionBarController and tests
The original CL to fix ActionBarControllerTest by using an
ActivityInstrumentationTestCase was too heavyweight and caused flaky
test. It was also indicative of a flaw in the original design that
ActionBar couldn't be mocked out easily.

Refactor the code so that:
1) getActionBar() is no longer needed
2) Respect the invariant that all access to the actionBar goes through the
ActionBarController (see the changes in ListsFragment) so that any change to
the ActionBar's hide offset can be tracked appropriately.

Bug: 17487016
Change-Id: I935b683ddb301ca5d3952889af1381bdbdcc97d3