Updating "Add Contact" to pre-populate full contact info.
1. For recent call entries with a name and phone number, the name will
now also be added.
2. For recent call entries which have complete address information provided
by nearby places, all information except the contact photo will be added.  The contact photo is
excluded, as the ContactsLoader would attempt to load the contact photo
while parsing the contact information.  On a slow connection this could
cause a noticeable pause while adding the contact.

Bug: 17308163
Change-Id: If2e78bd1257096b344ff2dd266dcdbaeed4f1471