Ignore all voicemails in the call log marked "deleted".
With the addition of the visual voicemail sync adapter, we want to have
a flag in the voicemail provider that will indicate that a voicemail was
deleted but not yet synced to the server. However, we no longer want
the entry to show up in the dialer call log, so make sure to ignore all
voicemails marked with "deleted" when querying the call log table.

Bug: 19236241
Change-Id: Ib80537c3b8630dd99b5e2b9b20d81015eafd114c