Remove expand/collapse animations in the Call Log.
- Remove animation logic in CallLogFragment.
- Remove expand listener from CallLogAdapter.
+ Update tests and constructor.

This preemptively removes expand/collapse behavior. While this
will result in a janky short-term experience, this is an incremental
step in shifting to RecyclerView and the new "Card" UI. I figured it
would be most efficient to delete it straight up, rather than coding
in intermediate logic to work with the RecyclerView because it will be
made obsolete by the Cards UI anyways.

In the long run, expand/collapse animation behavior will be
restored once the call log is migrated to RecyclerView and converted
to Cards.

Bug: 19372817
Change-Id: Icbdc7dcb9f2f2223456c29334c826d38917b087a