Simplify and rewire in-call buttons
* Fix bug where the switch camera option was never shown
* Fix bug where dialpad button was shown for video calls

* Move management of call button's actual visiblity into the
CallCardFragment. The CallCardPresenter is in charge of deciding
what functionality is available, but the CallCardFragment manages
the actual UI that is displayed.

* Simplify code that collapses call button functionality into
the overflow button. Any buttons that would cause the total number
of buttons to exceed the given threshold will be collapsed into
an overflow menu.

* Dynamically toggle the contentDescription of the hold button. This
is used for both accessibility as well as generating the menu option
if the button is collapsed into the overflow menu

* Removed some now unused strings and code.

Bug: 20127747
Bug: 20127925
Bug: 20090502

Change-Id: I33fda7ad74af926d8c3a507bff8f3c69c19ea57e