Video call upgrade/dowgrade request changes.
- fixed potential NPE in VideoCallFragment when setting preview size.
- moved photo load into the postExecute for the async task -- it is
already threaded and I was seeing intermittent concurrency issues.

- Changed CallButtonFragment to retrieve max # of buttons from config.xml.
  - Added override for wider screens (e.g. N6 and wider) to show an extra
    button.
- Reorganized call buttons so that the "Camera on/off" button is adjacent
to the flip camera button.
- Changed answer Glowpad to pass correct video state so that accepting a
video request uses the correct state (important for accepting requests to
turn camera back on).
- added new Session modification state REQUEST_REJECTED for when the remote
user explicitly declines the request.  This is used to trigger a
"video request rejected" message when the remote party rejects the
request.

Bug: 20257400
Change-Id: Ibe25eb045ee868748f91bf411f285629d36ebcd2