Cache the non-rounded large bitmap instead of the rounded version
Because a new copy of a rounded bitmap is created everytime,
the check against the cached copies of information was failing to
detect identical notifications, causing multiple full-screen
notifications to be sent to the notification manager.

This causes a delay in the time it takes for the HUN to be
hidden when accepting a call.

Bug: 20764430
Change-Id: I4f62d5091b7fc216312a612181a91e3ec4828661