Fix issues with fullscreen video call mode and incoming calls.
In onStateChange, when there is an incoming call, the primary call is
kept as the active call.  This is problematic for the logic in the
maybeAutoEnterFullscreen method which relied on the primary call to
determine if the UX should automatically switch to fullscreen mode.

Added code to ensure we auto-exit fullscreen mode based on the current
call context.

Bug: 20915113
Change-Id: If11c358c83c92a308d6f78f229354f7feb5bfea4