Fix issue where video surfaces show when dowgrading to audio in landscape.
The issue is the "inflateVideoCallViews" call in onViewCreated.  We should
not call inflateVideoCallViews unless we know that the call is a video
call.  Removing the call to inflateVideoCallViews in onViewCreated solves
the problem.  Once the UI sets up, onPrimaryCallChanged in the
VideoCallPresenter is called to establish the primary call.  At this point
inflateVideoCallViews will get called anyways, so having it done in the
onViewCreated is redundant.

Bug: 20188274
Change-Id: Iface049e28192d58cf85347f6e88ca6e38a6a9aa