Fix for glowpad not showing when dialing out + call waiting
The AnswerPresenter is responsible for initiating the creation
of the UI during a new incoming call, so there is no need
to do a null check on getUi in onIncomingCall.

Original fix was made in ag/651766 but seems to have been
lost in a series of merges.

Also, if the same instance of InCallPresenter is reused across
activity teardown and setups, this can result in mAnswerPresenter
getting lost from the set of incoming call listeners. To fix this,
always readd it at setup.

Bug: 22357229
Change-Id: Iecd80fa0187d756181ada7631c4246ad48f7131e