Add permission prompts for contacts and dialpad search
Update the following fragments to handle denied permissions
-Contacts Search (Contacts and Location)
-Dialpad Search (Phone)

Tweak and remove some of the onTouch listener logic as they are
no longer valid with the new UI. Instead of intercepting the touches
when the query is empty and returning to the main dialer activity,
allow the fragments to remain on screen if the permission request
UI is showing.

Modify signature of onEmptyViewActionButtonClicked to remove unused
permissions parameter.

Bug: 22174668
Change-Id: I96d00f2ab45df936dca602ac025f723638ac02c4