Last forwarded number and incoming call subject.
Last forwarded number:
- Added assets for the "forward" icon for last forwarded number.
- Modified InCall Call to store last forwarded number; also calls
onLastForwardedNumberChange callback (the last forwarded number is
received via a supp service update, so may change after call starts).

Call subject (i.e. instant lettering):
- Added placeholder subject_bubble asset to form the chat bubble for
incoming calls with a subject.
- Modified InCall Call to store the call subject (expected to be populated
in extras at start of call.
- Added code to hide the call status (e.g. "incoming call via XYZ") line
and primary call label (e.g. a location "California", or the number type
"Mobile" for the number).  This was necessary to make room for the call
subject bubble, and is in line with the UX mocks.
- Change call subject text color to background color of call card (per
UX mocks)
- Modified call notification to show call subject if it is specified.

- Moved code to show HD icon into common method.

Bug: 22685114
Change-Id: I22d9dae16658490e3245cfdd9c936bb0584cd6db