Show snackbar to undo last deleted voicemail.
Snackbar will appear for 3 seconds during which the user can undo the
last deletion. The way it works is the snackbar appears for 3 seconds
and a delayed callback is set for 3 seconds after which the voicemail is
permanently deleted from the database. If a second (or third or fourth)
voicemail is deleted subsequently, the previous voicemails that were
waiting for the undo timeout are deleted immediately.

Bug: 22460745
Change-Id: I84b70994275975e4e3020321884d382cc87098dc