Display unread count for voicemail and update when vm is read.
Query for number of unread voicemails when dialer starts and update this
number when the user reads the voicemails.
Pass the tab icons into the ViewPagerTabs class: changed so that the raw
image is passed instead of the modified drawable because the ripple is
applied in through the ViewPagerTabs layout instead of through the
drawable containing the icons.

Bug: 21086059

Change-Id: Ie426b042d72926ea3feb7baa4717108776152754