Use Dialer utility for getting LookupUri.
The intention is to consolidate how InCall and Dialer do Contact
lookups, but it's actually hard to do this truly without more
refactoring because they have different model and query numbers.

Just get the same URI for the initial lookup now... specific
casing, such as in the secondary lookups if the initial SIP query
fails, is still implemented separately in ContactInfoHelper and
CallerInfo#doSecondaryLookupIfNecessary.

Bug: 19993243
Change-Id: I8c5fd75d3cda84607d5ee5858b2b73adf20eebb8